add touch response widget
